fix: remove table_owner retrieved from manifest
In DBT documentation, 2 owners are displayed: on the top line, we have the dbt owner as defined in the metadata, on the second line it is the owner of the sql representation of the model in Athena. But Athena does not support table owner except if used with LakeFormation. But in any case it is not an owner, in a traditional way as we can see it in Postgres. Since dbt 1.8 remove the ability to use the manifest object when retrieving catalog or schemas, we were forced to remove it. But it feels like we can remove it now because we have an owner displayed and the documentation and it reflects more what happened in reality.