chore: add tests for all on_schema_change modes
The previous commit included tests for the sync_all_columns mode, since that had a bug. There are no functional tests that cover the on_schema_change modes, so this adds coverage for the other three modes.

I also added an additional check to ensure the column names have actually been changed (or not changed).