fix: vacuum more runs needed error
When there is too many files to process during a vacuum, dbt model fails
with this error: "ICEBERG_VACUUM_MORE_RUNS_NEEDED: Removed 20000 files
in this round of vacuum, but there are more files remaining. Please
run another VACUUM command to process the remaining files."

We apply therefore the same logic as we did for the optimize. There
is also an attempt to gather the code since they have the same logic.